requirements.txt
scapy>=2.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
python-dateutil>=2.8.2
streamlit>=1.20
beautifulsoup4>=4.11.0
python-dotenv>=0.20.0